% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_gradient}
\alias{check_gradient}
\title{Check Analytic Gradient against Numeric Gradient}
\usage{
check_gradient(par, fn, gr, tol = 1e-04, ...)
}
\description{
Check Analytic Gradient against Numeric Gradient
}
\keyword{internal}
