% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{generate_starts}
\alias{generate_starts}
\title{Generate Multiple Starting Points}
\usage{
generate_starts(par, lower, upper, n_starts)
}
\description{
Generate Multiple Starting Points
}
\keyword{internal}
