% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_symbolic.R
\name{print.smart_sym}
\alias{print.smart_sym}
\title{Print Smart Symbolic objects}
\usage{
\method{print}{smart_sym}(x, ...)
}
\arguments{
\item{x}{The smart_sym object to print.}

\item{...}{Additional arguments.}
}
\description{
Prints the symbolic derivative nicely.
}
