% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_loglik.R
\name{smart_loglik}
\alias{smart_loglik}
\title{Fetch Known Negative Log-Likelihood Functions}
\usage{
smart_loglik(distn_name)
}
\arguments{
\item{distn_name}{A character string of the distribution name (e.g., "normal", "gumbel").}
}
\value{
An R function mathematically prepared for symbolic differentiation.
}
\description{
Automatically retrieves the mathematical formula for the Negative Log-Likelihood
of 30 common statistical distributions.
}
