% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_optim.R
\name{smart_optim}
\alias{smart_optim}
\title{Smart Optimization Engine}
\usage{
smart_optim(
  par,
  fn,
  gr = NULL,
  lower = -Inf,
  upper = Inf,
  method = "auto",
  n_starts = 1,
  scale = TRUE,
  check_grad = TRUE,
  explain = TRUE,
  benchmark = FALSE,
  control = list(),
  ...
)
}
\arguments{
\item{par}{Initial values for the parameters to be optimized over.}

\item{fn}{A function to be minimized (or maximized).}

\item{gr}{A function to return the gradient.}

\item{lower}{Bounds on the variables for the "L-BFGS-B" method.}

\item{upper}{Bounds on the variables for the "L-BFGS-B" method.}

\item{method}{The optimization method to be used.}

\item{n_starts}{Number of random starts to avoid local minima.}

\item{scale}{Scaling parameter for the variables.}

\item{check_grad}{Logical. Should the gradient be checked numerically?}

\item{explain}{Logical. Should the function print an explanation of the optimization process?}

\item{benchmark}{Logical. Should benchmarking be performed?}

\item{control}{A list of control parameters to pass to optim().}

\item{...}{Additional arguments to pass to the function \code{fn}.}
}
\description{
Finds the optimal parameters for a given function with advanced fallbacks and diagnostics.
}
