% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_symbolic.R
\name{smart_sym_deriv}
\alias{smart_sym_deriv}
\title{Smart Symbolic Differentiation}
\usage{
smart_sym_deriv(fn, var = "x")
}
\arguments{
\item{fn}{An R function to differentiate.}

\item{var}{Character vector of variables to differentiate with respect to (default is "x").}
}
\value{
A new R function representing the derivative.
}
\description{
Takes an R function and returns a brand new R function that represents its exact symbolic derivative.
}
